verdict/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bcem-verdict"
version = "0.1.0"
description = "BCEM Email Security Scanner — Verdict Worker"
requires-python = ">=3.10"
dependencies = [
    "celery[redis]>=5.3,<6",
    "redis>=5.0,<6",
    "httpx>=0.27,<1",
    "psycopg[binary]>=3.1,<4",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "respx>=0.21",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
